+++ source/ProcessManager.Components.ValueObjects/ActorRole.cs
@@ -14,9 +14,8 @@
 
 using System.Text.Json.Serialization;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects;
+namespace Energinet.DataHub.ProcessManager.Components.ValueObjects;
 
-[Serializable]
 public class ActorRole : DataHubType<ActorRole>
 {
     public static readonly ActorRole MeteringPointAdministrator = new("MeteringPointAdministrator", "DDZ");