+++ source/ProcessManager.Components.ValueObjects/DataHubType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects;
+namespace Energinet.DataHub.ProcessManager.Components.ValueObjects;
 
 /// <summary>
 /// Base class for enumeration types with a code (typically used for our domain object with CIM codes)