+++ source/ProcessManager.Orchestrations/Components/DataHub/Measurements/Mappers/MeteredDataToMeasurementMapper.cs
@@ -12,61 +12,62 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Measurements.Contracts;
+using MeasurementTypes=Energinet.DataHub.Measurements.Contracts;
+using PMTypes=Energinet.DataHub.ProcessManager.Components.ValueObjects;
 
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Components.DataHub.Measurements.Mappers;
 
 public static class MeteredDataToMeasurementMapper
 {
-    internal static Dictionary<ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Quality, Energinet.DataHub.Measurements.Contracts.Quality> Quality { get; } = new()
+    internal static Dictionary<PMTypes.Quality, MeasurementTypes.Quality> Quality { get; } = new()
     {
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Quality.NotAvailable, Energinet.DataHub.Measurements.Contracts.Quality.QMissing },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Quality.Estimated, Energinet.DataHub.Measurements.Contracts.Quality.QEstimated },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Quality.AsProvided, Energinet.DataHub.Measurements.Contracts.Quality.QMeasured },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Quality.Calculated, Energinet.DataHub.Measurements.Contracts.Quality.QCalculated },
+        { PMTypes.Quality.NotAvailable, MeasurementTypes.Quality.QMissing },
+        { PMTypes.Quality.Estimated, MeasurementTypes.Quality.QEstimated },
+        { PMTypes.Quality.AsProvided, MeasurementTypes.Quality.QMeasured },
+        { PMTypes.Quality.Calculated, MeasurementTypes.Quality.QCalculated },
     };
 
-    internal static Dictionary<ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Resolution, Resolution> Resolution { get; } = new()
+    internal static Dictionary<PMTypes.Resolution, MeasurementTypes.Resolution> Resolution { get; } = new()
     {
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Resolution.QuarterHourly, Energinet.DataHub.Measurements.Contracts.Resolution.RPt15M },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.Resolution.Hourly, Energinet.DataHub.Measurements.Contracts.Resolution.RPt1H },
+        { PMTypes.Resolution.QuarterHourly, MeasurementTypes.Resolution.RPt15M },
+        { PMTypes.Resolution.Hourly, MeasurementTypes.Resolution.RPt1H },
     };
 
-    internal static Dictionary<ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit, Energinet.DataHub.Measurements.Contracts.Unit> MeasurementUnit { get; } = new()
+    internal static Dictionary<PMTypes.MeasurementUnit, MeasurementTypes.Unit> MeasurementUnit { get; } = new()
     {
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.KilowattHour, Unit.UKwh },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.MegawattHour, Unit.UMwh },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.MegaVoltAmpereReactivePower, Unit.UMvar },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.KiloVoltAmpereReactiveHour, Unit.UKvarh },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.Kilowatt, Unit.UKw },
-        { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeasurementUnit.MetricTon, Unit.UTonne },
+        { PMTypes.MeasurementUnit.KilowattHour, MeasurementTypes.Unit.UKwh },
+        { PMTypes.MeasurementUnit.MegawattHour, MeasurementTypes.Unit.UMwh },
+        { PMTypes.MeasurementUnit.MegaVoltAmpereReactivePower, MeasurementTypes.Unit.UMvar },
+        { PMTypes.MeasurementUnit.KiloVoltAmpereReactiveHour, MeasurementTypes.Unit.UKvarh },
+        { PMTypes.MeasurementUnit.Kilowatt, MeasurementTypes.Unit.UKw },
+        { PMTypes.MeasurementUnit.MetricTon, MeasurementTypes.Unit.UTonne },
     };
 
-    internal static Dictionary<ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType, Energinet.DataHub.Measurements.Contracts.MeteringPointType> MeteringPointType { get; } = new()
+    internal static Dictionary<PMTypes.MeteringPointType, MeasurementTypes.MeteringPointType> MeteringPointType { get; } = new()
         {
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.Consumption, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptConsumption },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.Production, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.Exchange, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptExchange },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.VeProduction, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptVeProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.Analysis, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptAnalysis },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NotUsed, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNotUsed },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.SurplusProductionGroup6, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptSurplusProductionGroup6 },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NetProduction, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNetProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.SupplyToGrid, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptSupplyToGrid },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.ConsumptionFromGrid, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptConsumptionFromGrid },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.WholesaleServicesInformation, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptWholesaleServicesInformation },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.OwnProduction, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptOwnProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NetFromGrid, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNetFromGrid },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NetToGrid, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNetToGrid },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.TotalConsumption, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptTotalConsumption },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NetLossCorrection, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNetLossCorrection },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.ElectricalHeating, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptElectricalHeating },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.NetConsumption, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptNetConsumption },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.OtherConsumption, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptOtherConsumption },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.OtherProduction, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptOtherProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.CapacitySettlement, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptEffectPayment },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.ExchangeReactiveEnergy, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptExchangeReactiveEnergy },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.CollectiveNetProduction, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptCollectiveNetProduction },
-            { ProcessManager.Orchestrations.Abstractions.Components.Datahub.ValueObjects.MeteringPointType.CollectiveNetConsumption, Energinet.DataHub.Measurements.Contracts.MeteringPointType.MptCollectiveNetConsumption },
+            { PMTypes.MeteringPointType.Consumption, MeasurementTypes.MeteringPointType.MptConsumption },
+            { PMTypes.MeteringPointType.Production, MeasurementTypes.MeteringPointType.MptProduction },
+            { PMTypes.MeteringPointType.Exchange, MeasurementTypes.MeteringPointType.MptExchange },
+            { PMTypes.MeteringPointType.VeProduction, MeasurementTypes.MeteringPointType.MptVeProduction },
+            { PMTypes.MeteringPointType.Analysis, MeasurementTypes.MeteringPointType.MptAnalysis },
+            { PMTypes.MeteringPointType.NotUsed, MeasurementTypes.MeteringPointType.MptNotUsed },
+            { PMTypes.MeteringPointType.SurplusProductionGroup6, MeasurementTypes.MeteringPointType.MptSurplusProductionGroup6 },
+            { PMTypes.MeteringPointType.NetProduction, MeasurementTypes.MeteringPointType.MptNetProduction },
+            { PMTypes.MeteringPointType.SupplyToGrid, MeasurementTypes.MeteringPointType.MptSupplyToGrid },
+            { PMTypes.MeteringPointType.ConsumptionFromGrid, MeasurementTypes.MeteringPointType.MptConsumptionFromGrid },
+            { PMTypes.MeteringPointType.WholesaleServicesInformation, MeasurementTypes.MeteringPointType.MptWholesaleServicesInformation },
+            { PMTypes.MeteringPointType.OwnProduction, MeasurementTypes.MeteringPointType.MptOwnProduction },
+            { PMTypes.MeteringPointType.NetFromGrid, MeasurementTypes.MeteringPointType.MptNetFromGrid },
+            { PMTypes.MeteringPointType.NetToGrid, MeasurementTypes.MeteringPointType.MptNetToGrid },
+            { PMTypes.MeteringPointType.TotalConsumption, MeasurementTypes.MeteringPointType.MptTotalConsumption },
+            { PMTypes.MeteringPointType.NetLossCorrection, MeasurementTypes.MeteringPointType.MptNetLossCorrection },
+            { PMTypes.MeteringPointType.ElectricalHeating, MeasurementTypes.MeteringPointType.MptElectricalHeating },
+            { PMTypes.MeteringPointType.NetConsumption, MeasurementTypes.MeteringPointType.MptNetConsumption },
+            { PMTypes.MeteringPointType.OtherConsumption, MeasurementTypes.MeteringPointType.MptOtherConsumption },
+            { PMTypes.MeteringPointType.OtherProduction, MeasurementTypes.MeteringPointType.MptOtherProduction },
+            { PMTypes.MeteringPointType.CapacitySettlement, MeasurementTypes.MeteringPointType.MptEffectPayment },
+            { PMTypes.MeteringPointType.ExchangeReactiveEnergy, MeasurementTypes.MeteringPointType.MptExchangeReactiveEnergy },
+            { PMTypes.MeteringPointType.CollectiveNetProduction, MeasurementTypes.MeteringPointType.MptCollectiveNetProduction },
+            { PMTypes.MeteringPointType.CollectiveNetConsumption, MeasurementTypes.MeteringPointType.MptCollectiveNetConsumption },
         };
 }