+++ docs/ProcessManager.Client/ReleaseNotes/ReleaseNotes.md
@@ -7,6 +7,10 @@
     - TransactionId
     - MeteringPointId
 
+## Version 0.27.1
+
+- No functional changes.
+
 ## Version 0.27.0
 
 - Update Nuget package `Grpc.Tools`